Refactor UpdateParkingAreaAsync to use DTO from request body
Refactored the `UpdateParkingAreaAsync` method in `ParkingAreaController.cs` to simplify parameter handling by removing the `id` parameter from the route and requiring the `UpdateParkingAreaReqDto` object to contain the `ParkingAreaId`. Updated the `HttpPut` attribute accordingly. Added a `[Required]` attribute to the `ParkingAreaId` property in `UpdateParkingAreaReqDto` to ensure it is always provided. Removed redundant manual property mapping, directly using the DTO object for the update operation.